# Track PostingsForMatchers cached promise completion timestamp in trace